requirements.txt
fastapi==0.92.0
typing==3.7.4.3
starlette==0.25.0
httpx==0.23.3
aiohttp==3.8.3
uuid==1.30
uvicorn==0.20.0
python-dotenv==1.0.0